NODE-308: Merged DownloadManagerImpl and DownloadManager into one file.
